Restructured clients directory
