+++ clients/ibm-pi-vpn_integration_test.go
@@ -6,7 +6,7 @@ import (
 
 	utl "internal/testutils"
 
-	client "github.com/IBM-Cloud/power-go-client/clients/instance"
+	client "github.com/IBM-Cloud/power-go-client/clients"
 	"github.com/IBM-Cloud/power-go-client/power/models"
 	"github.com/stretchr/testify/require"
 )
