+++ examples/images/main.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"log"
 
-	v "github.com/IBM-Cloud/power-go-client/clients/instance"
+	v "github.com/IBM-Cloud/power-go-client/clients"
 	ps "github.com/IBM-Cloud/power-go-client/ibmpisession"
 	"github.com/IBM-Cloud/power-go-client/power/models"
 	"github.com/IBM/go-sdk-core/v5/core"
